Change serialization for int and float to work with objects
It is good to have opportunity to pass objects which have valueOf() method in resolvers of Int and Float types.For example, I have such objects as "Money", "Price" and so on in my app, and they have valueOf() method which return value. Now I need to call valueof method in every resolver of such field or write custom scalar or directive.I think changes I did  are implied in native type behavior of Int and Float fields.